Move check in using loan replacement tests to separate class CIRC-154
